+++ activemodel/CHANGELOG.md
@@ -1,3 +1,23 @@
+*   Ability to specify multiple contexts when defining a validation.
+
+    Example:
+
+        class Person
+          include ActiveModel::Validations
+
+          attr_reader :name
+          validates_presence_of :name, on: [:verify, :approve]
+        end
+
+        person = Person.new
+        person.valid?                           # => true
+        person.valid?(:verify)                  # => false
+        person.errors.full_messages_for(:name)  # => ["Name can't be blank"]
+        person.valid?(:approve)                 # => false
+        person.errors.full_messages_for(:name)  # => ["Name can't be blank"]
+
+    *Vince Puzzella*
+
 *   `attribute_changed?` now accepts parameters which check the old and new value of the attribute
 
     `model.name_changed?(from: "Pete", to: "Ringo")`